Linux: add os-release support
Python's platform.linux_distribution() returns an empty tuple if a Linux
distribution is not part of the '_supported_dists' list.

Add a method to parse the os-release file, the new standard release file,
which replaces the multitude of per-distribution ones.

Also, platform.linux_distribution() is being deprecated and then removed
in future Python releases.