Avoid crashing if LC_ALL is not defined by user
When LC_ALL is not defined in the user environment (which is the case when I connect via SSH to both an Ubuntu and Raspbian OS), then launching glances result in an error that the '' locale setting is unsupported (line 98 in __init.py__ fails).
I've added a try/except statement and if locale.Error is raised simply ignore it. It means that glances will be running with the C locale. But it is better than not running at all. :-)