Big refactoring of the Web UI
Move logic from controller and views to the new services GlancesStats and GlancesPlugins : GlanceStats load data from api and set data in plugins. Plugins are responsible for transforming data for the view.
GetAlert is now in GlancesPlugin
Use view data instead of getAlert when is possible
Remove directives for monitor plugin