Fix UTF-8 characters being output on latin1 terminals
This fix translates the characters to an appropriate encoding for the
terminal. "Â°C" (A acute, degrees, C) was being output on latin1
terminals, which is a bug since the LC_CTYPE locale setting should tell
the program whether it's okay to output Unicode.

This follows the recommendation from the official curses library
documentation: https://docs.python.org/2/library/curses.html

Steps to reproduce the issue:

1. Install pysensors, enable any kernel modules necessary to support it
2. Run `LANG=en_US rxvt -e glances`
3. Look in SENSORS section

The corrupt Unicode character sequence will appear.