exports: riemann export implemented and documented
This patch introduces a new export method for sending metrics to
__Riemann__ server using _TCP_ protocol. The module can be enabled like
all the other exports in __Glances__, specifying the `--export-riemann`
option.

Sending metrics to Riemann requires
[bernhard](https://github.com/banjiewen/bernhard) that is
the most popular _Riemann client_ written in python.

In the configuration file you can only specify the `host` and `port`
that riemann is listening for new connections on. Later new options are
planned to be added like __SSL__ support and __UDP__.

Documentation and manpages are refreshed according to the modifications
in the config options and introducing the usage of Riemann export.

Signed-off-by: Gergő Nagy <grigori.grant@gmail.com>