Implement custom snap version string generation
The snap version string generated by the `git` value of the `version`
key isn't mimicing the previous versioning style due to the prefixing
`v` in the git tag, this patch replace it with a custom implementation
based on the `version-script` key.

Note the `version` key is still required to exist due to schema
requirements, but has no effect when the `version-script` key is
defined.

Reference:

    Snapcraft version scripts - snapcraft - snapcraft.io
    https://forum.snapcraft.io/t/snapcraft-version-scripts/106

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>